GO
The present REPORT has introduced an innovative, efficient and real-time cost beneficial technique that enables
user to hear the contents of text images instead of reading through them. It combines the concept of Optical
Character Recognition (OCR) and Text to Speech Synthesizer (TTS) in Raspberry pi. This kind of system helps
visually impaired people to interact with computers effectively through vocal interface. Text Extraction from color
images is a challenging task in computer vision. Text-to-Speech conversion is a method that scans and reads English
alphabets and numbers that are in the image using OCR technique and changing it to voices. This report describes
the design, implementation and experimental results of the device. Here we have arranged a series of experiment as
code of data that will be mined using virtual mp3 sound-track, also other uses of text-to-speech mining as scanner